relocate techSingleTech
