+++ src/App.tsx
@@ -6,9 +6,9 @@ import { ruRU } from '@mui/x-data-grid';
 import Header from './components/header/Header';
 import RequireAuth from './components/requireAuth.tsx/RequireAuth';
 import RadarConstructor from './pages/admin/radarConstructor/RadarConstructor';
-import TechSinglePage from './pages/admin/techSinglePage/TechSinglePage';
 import Constructor from './pages/constructor/Constructor';
 import Main from './pages/main/Main';
+import TechSinglePage from './pages/techSinglePage/TechSinglePage';
 import TechRadar from './pages/techradar/TechRadar';
 import { ConstructorMode } from './store/editRadarSlice';
 import { useCredentials } from './store/hooks';
